refactor(gabinet): split detailed-data-tab into section sub-components (closes #4459)
Extract the five distinct sections of the ~1000-line DetailedDataTab into
focused sub-components under sections/:

- PersonalInfoSection — personal info + photo upload
- EmploymentSection — employment type, role, dates, notes
- QualificationsSection — skills, experience, certifications
- AssignedTreatmentsSection — treatment picker dialog
- CompensationSection — salary, commission, bank account

Shared utilities extracted to section-header.tsx (SectionHeader, ReadOnlyField)
and types.ts (EmployeeFormData, EMPLOYMENT_TYPES, EmploymentType).

The parent DetailedDataTab retains all state and handlers; sections receive
only the props they need and render their own read/edit views.
